+++ lang/src/main/kotlin/org/partiql/lang/planner/transforms/optimizations/ConcatWindowFunction.kt
@@ -4,7 +4,6 @@ import org.partiql.lang.domains.PartiqlPhysical
 import org.partiql.lang.errors.ProblemHandler
 import org.partiql.lang.planner.PartiQLPhysicalPass
 
-// TODO: Remove from experimental once https://github.com/partiql/partiql-docs/issues/31 is resolved and a RFC is approved
 internal fun createConcatWindowFunctionPass(): PartiQLPhysicalPass =
     ConcatWindowFunction()