+++ lang/src/test/kotlin/org/partiql/lang/util/testdsl/IonResultTestCase.kt
@@ -2,7 +2,6 @@ package org.partiql.lang.util.testdsl
 
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
-import org.partiql.annotations.PartiQLExperimental
 import org.partiql.lang.CompilerPipeline
 import org.partiql.lang.ION
 import org.partiql.lang.eval.CompileOptions
@@ -69,7 +68,6 @@ data class IonResultTestCase(
         }
 }
 
-@OptIn(PartiQLExperimental::class)
 internal fun IonResultTestCase.runTestCase(
     db: MockDb,
     target: EvaluatorTestTarget,